refactor(gateway): GAR-439 slice 9.a — extract admin/shared.rs
TDD REFACTOR puro (zero behavior change). Slice 9.a do epic
GAR-430 / Q9 (Quality Gates Phase 3.6). Primeira extração do
refactor de admin/handlers.rs (3300 LOC) em 6 PRs pequenos.

Movido para o novo `admin/shared.rs`:
- pub struct AdminState (8 LOC com #[derive(Clone)])
- pub fn derive_encryption_key() (51 LOC, lógica PBKDF2 +
  fallback file-or-generate)

Re-export em handlers.rs (`pub use super::shared::{AdminState,
derive_encryption_key};`) preserva paths externos
`super::handlers::AdminState` e `super::handlers::derive_encryption_key`
consumidos por admin/routes.rs:9,16. Zero alteração em routes.rs.

Imports limpos em handlers.rs:
- Removidos: std::sync::Arc, tokio::sync::Mutex,
  super::store::AdminStore, crate::state::SharedState
  (usados apenas por AdminState que saiu)
- Mantidos: ring::rand::{SecureRandom, SystemRandom} (ainda
  usados por encrypt_value/decrypt_value que ficam em handlers.rs
  até slice 9.c).

Validação local:
- cargo fmt --check ✓
- cargo check -p garraia-gateway --locked ✓
- cargo +1.92 check --workspace --exclude garraia-desktop --locked ✓
- cargo clippy -p garraia-gateway --all-targets -- -D warnings ✓
- cargo test -p garraia-gateway --lib admin:: → 23/23 passed

LOC handlers.rs: 3300 → 3245 (-55)
LOC shared.rs: novo, 81 linhas
LOC mod.rs: +1 (`pub mod shared;`)

Reviews paralelos (4):
- code-reviewer: APPROVED, 0 blockers, 1 LOW herdada (NonZeroU32
  unwrap pre-existente)
- security-auditor: PASS no slice; 2 MEDIUM PRE-EXISTENTES (salt
  fixo, PBKDF2 100k < 210k NIST 2023) e 2 LOW (dual public path,
  master.key sem chmod 0600). Nenhum introduzido por este slice.
  Será aberta sub-issue de hardening como follow-up.
- rust-refactor-engineer: PASS, byte-equivalence confirmada
- privacy-hygiene-reviewer: CLEAN
