+++ crates/garraia-gateway/src/admin/mod.rs
@@ -3,4 +3,5 @@ pub mod handlers;
 pub mod middleware;
 pub mod rbac;
 pub mod routes;
+pub mod shared;
 pub mod store;
